requirements.txt
Flask
Gunicorn
line-bot-sdk
firebase-admin